Implemented the function of requesting subgroups to see visualization of subgroups. For example, if the user requests "all", then all 8 races would show up in the bar graph. If the user requests "White Male, White Female", then only these two requested group would show up in the bar graph.
Still need to figure out how many combinations we want to provide for the user.
Since eight factorial of combination is hard to implement with if-else statement.
Or if anyone has better way to implement the subgroup requests thing, we could try.

Last thing, I don't know how to change the bar colors...